imx-base-extend.inc: Use ??= for PREFERRED_VERSION overrides [YOCIMX-6362]
The user cannot override the PREFERRED_VERSION settings in
imx-base-extend.inc because machine configuration takes precedence over
layer.conf or local.conf.

The fix starts with imx-base.inc in meta-freescale:

https://github.com/Freescale/meta-freescale/commit/5b47898ff939791ab426fff17821aca3476ddd7a

```
The variable values set in meta-freescale with a default value, i.e.
with ?=, can be overridden just once by a user, with =.

Change the PREFERRED_VERSION settings to a weak default value, i.e.
with ??=, so users of meta-freescale are not so limited. When a
variable value is set with a weak default value, the variable can be
overridden by any successively higher priority layer with ??=.
```

Use ??= here as well so the user can override.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>