imx-base-extend.inc: Fix image type for initramfs [YOCIMX-6055]
The fix to add .zst images broke the initramfs build. The design is
fragile such that the override to specialize the image types with
the .zst images cannot work because it also overrides the override in
mfgtool-initramfs-image.bbclass.

Fix the problem by removing the new overrides and adding the .zst
images to an existing override using a variable.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>