u-boot-imx-tools: Fix target and SDK builds [YOCIMX-6531]
The class-target and class-nativesdk builds are broken after the
upgrade to 2022.04, with failures due to missing headers uuid.h and
gnutls.h, and another failure trying to execute the locally built tool
tools/bmp_logo. These failures are fixed in OE-Core master.

https://github.com/openembedded/openembedded-core/commit/a9c5f9033343ad2e26539854ccde2f116528f235

Since OE-Core kirkstone is on 2022.01, backport the fix.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>