imx-gpu-viv: Fix package removal error
The package removal logic in the install has a logical error where the
for loop with wildcards can match and resolve to paths on the build
machine. This is incorrect, as the paths should only be checked
relative to the install folder. Fix the problem by disabling
pathname expansion for the for loop command.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>