imx-atf: Fix BOOT_TOOLS workaround for Cockpit
The Cockpit build no longer works, failing in do_deploy:

```
| cp: cannot stat '.../deploy-imx-atf/bl31-imx8qm.bin': No such file or directory
```

The problem started with meta-freescale recipe dropping the custom
TF-A folder [1]. We still depend on the custom TF-A folder, so the
folder usage was restored with a do_deploy:append(). This function
fails when used with the override do_deploy:imx8qm-cockpit().

Rework the solution by moving the custom folder handling directly into
do_deploy() and extending the override do_deploy:imx8qm-cockpit() with
the corresponding custom folder handling.

[1] https://github.com/Freescale/meta-freescale/commit/1ec9540091c6d2cb7faf679f3b01916bfb0a0397

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>