weston-10: Rename weston 10 recipe to weston-10 [YOCIMX-6716]
Weston 11 drops fbdev support, which is required for i.MX 6 & 7, so the
older weston 10 recipe is maintained. However, our build infrastructure,
external to Yocto, does not support a single package with multiple
versions.

So, to facilitate maintenance of 2 weston recipes, rename the weston 10
recipe to weston-10.

Note that the files that are shared between the two recipes are moved
to the common `files` folder.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>