config: Disable upstream recipe kernel-module-nxp89xx
The build kernel-module-nxp89xx is failing. This package should not be
building because it is renamed [YOCIMX-6690]. The fix in
imx-base-extend.inc does no longer work because the old name is now
added via MACHINE_EXTRA_RRECOMMENDS:

https://github.com/Freescale/meta-freescale/commit/029d698fe9523f88423241fcc6828f1ddc405045

Adjust the fix by dropping the change in imx-base-extend.inc and
co-locating the upstream override in layer.conf.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>