vulkan-samples: Switch to fork prabhusundar [MGS-6998]
Use the fork prabhusundar, which picks up the latest fixes plus an update
of glfw to support XDG shell.

It's okay to use the latest vulkan-samples even though Vivante GPU drivers are
on Vulkan 1.2.

Drop the build patch 0005-glfw-fix-glfw3.h-gl-error.patch, it is no longer
needed.

Drop the dependency on extra-cmake-modules, it is no longer needed.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>