gstreamer1.0-plugins-good: Drop libsoup overrides [LF-8385]
The FB and Wayland distro builds are failing in do_rootfs:
```
The following packages have unmet dependencies:
gstreamer1.0-plugins-good-soup : Depends: libsoup-2.4 but it is not installable
E: Unable to correct problems, you have held broken packages.
```

The problem is the recipe contains libsoup overrides that were
appropriate for 1.20.3.imx but not for the upgrade to 1.22.0.imx.

Drop the libsoup overrides since they are no longer needed.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>