distro: Hide meta-freescale gstreamer
gstreamer builds are failing:

```
WARNING: preferred version 1.22.0.imx of gstreamer1.0-plugins-base not available (for item gstreamer1.0-plugins-base-videoconvertscale)
ERROR: Multiple versions of gstreamer1.0-plugins-base are due to be built (/langdale/sources/meta-imx/meta-bsp/recipes-multimedia/gstreamer/gstreamer1.0-plugins-base_1.22.0.imx.bb /langdale/sources/meta-freescale/recipes-multimedia/gstreamer/gstreamer1.0-plugins-base_1.20.3.imx.bb). Only one version of a given PN should be built in any given build. You likely need to set PREFERRED_VERSION_gstreamer1.0-plugins-base to select the correct version or don't depend on multiple versions.
```

The problem seems to be with this change in meta-freescale [1], a
workaround to make 1.20.3.imx recipes compatible with the 1.22 change
to consolidate videoconvert and videoscale into one package
videoconvertscale.

[1] https://github.com/Freescale/meta-freescale/commit/2a9cad5187885a5de72ebedf4a9512c1f761c654

The gstreamer recipes use dynamic packages. Without the change, a log
shows that 1.22.0.imx is selected for the dynamic package
videoconvertscale. With the change, the explicit package in 1.20.3.imx
is selected instead.

This doesn't fully explain the behavior, but since the gstreamer
1.20.3.imx recipes in meta-freescale are not needed, the behavior can
be avoided simply with a BBMASK of the folder.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>