qtmultimedia: Remove ffmpeg [YOCIMX-6921]
For Qt 6.5, ffmpeg is added to the default qtmultimedia configuration.
The package is commercial, so remove it if commercial is not accepted.

Fixes:
```
ERROR: Nothing PROVIDES 'ffmpeg' (but sources/meta-qt6/recipes-qt/qt6/qtmultimedia_git.bb DEPENDS on or otherwise requires it)
ffmpeg was skipped: because it has a restricted license 'commercial'. Which is not listed in LICENSE_FLAGS_ACCEPTED
```

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>