optee-os: Fix tee build for cockpit [LF-8867]
The build is creating the same tee.bin for a72 and a53. The problem is
that the base recipe sets PLATFORM=imx-mx8qmmek, so both builds are
generating for this PLATFORM. Override PLATFORM instead of
PLATFORM_FLAVOR to fix this.

The deploy folder file tee.bin is linked to the poorly named
tee.mx8qmmek.bin. Fix this by overriding OPTEEMACHINE with the platform
for the a72.

Also replace objcopy with a simply copy of tee-raw.bin.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>