optee-os: Fix tee build for cockpit [LF-8727]
The build is creating the same tee.bin for a72 and a53. The problem is
that the base recipe sets PLATFORM=imx-mx8qmmek, so both builds are
generating for this PLATFORM. Override PLATFORM instead of
PLATFORM_FLAVOR to fix this.

The deploy folder file tee.bin is linked to the poorly named
tee.mx8qmmek.bin. Fix this by overriding PLATFORM_FLAVOR with the platform
for the a72.

Also replace objcopy with a simply copy of tee-raw.bin.

This commit is cherry-picked from 6.1 release and is adjusted because
the 5.15 release does not use meta-arm for optee.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
(cherry picked from commit ab1bf86d8bf25c8e3e43bcf5de33783faf0303d2)