packagegroup-fsl-tools-gpu-external: Remove kmscube from 8M Mini
There is a build break for 8M Mini:
```
ERROR: Nothing PROVIDES 'virtual/libgles3' (but sources/poky/meta/recipes-graphics/kmscube/kmscube_git.bb DEPENDS on or otherwise requires it). Close matches:
```

This results because of a fix to the dependencies of kmscube:

https://github.com/openembedded/openembedded-core/commit/63f338255000874c62bf01ccc11f2c100014c3d9

Fix the break by removing kmscube for 8M Mini.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>