qtwebengine: Add xdamage dependency [YOCIMX-7428]
The build is creating empty packages, resulting in a runtime error:
```
./quicknanobrowser: error while loading shared libraries: libQt6WebEngineQuick.so.6: cannot open shared object file: No such file or directory
```

The problem is that the build is actually failing, but the failure
is not being passed to bitbake:
```
QtWebEngine will not be built: Could not find all necessary libraries for qpa-xcb support. X11:YES LIBDRM:YES XCOMPOSITE:NO XCURSOR:NO XRANDR:NO XI:NO XPROTO:YES XSHMFENCE:YES XTST:NO
```

This started after OpenGL was removed from imx-gpu-viv. That change
resulted in a build break for qtwebengine:
```
| ERROR: Feature "webengine_webrtc": Forcing to "ON" breaks its condition:
|     XDAMAGE_FOUND OR NOT QT_FEATURE_webengine_ozone_x11
| Condition values dump:
|     XDAMAGE_FOUND = ""
|     QT_FEATURE_webengine_ozone_x11 = "ON"
```

The build break was 'fixed' in 9b8e1f5bbb by removing x11 support, but
this only resulted in the empty package and runtime error described
above.

Restore the x11 support and fix the original build break by adding an
xdamage dependency.

Fixes: 9b8e1f5bbb
Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>