openssh: Remove the bbappend
The bbappend is added to disable rng-tools originally as [1], but the
rng-tools has been already removed in oe-core as [2]. So no need to
keep this bbappend which used to disable rng-tools. And the means used
to disable rng-tools [1] also make the sshd service not work after the
logic [3] introduced.

So remove the useless bbappend.

[1] 6c9ad14
[2] https://git.openembedded.org/openembedded-core/commit/?id=868dfb46d96a27ec9041cb902fb769330277257d
[3] https://git.openembedded.org/openembedded-core/commit/?id=bc830ad3c6a11af1a350dca7f33f0682aeee0d21

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
Signed-off-by: Mingli Yu <mingli.yu@windriver.com>