imx-setup-release.sh: Fix error when no DISTRO is given
If the setup script is executed for MACHINE=imx8* but with no DISTRO,
an error is reported:
```
 ERROR - Only Wayland distros are supported for i.MX 8 or i.MX 8M
```

The default distro is fsl-imx-xwayland, so this is wrong. The problem
is the test that reports this error checks the value of DISTRO, when
in fact it is the variable FSLDISTRO that is set.

Also, drop the if-else call to the setup program since both cases
are identical.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>