dpdk: Update dpdk_22.11.bb to commit 4d94e5c
The latest source uses shared libs and adds the ability to install
the examples. Adapt the recipe to these changes.

By default, the source installs the implementation libs in a
sub-folder of /usr/lib. This is non-standard and is fixed by clearing
the variable `drivers_install_subdir`.

By default, the source installs the source code for all the examples in
/usr/share/dpdk. This is non-standard and is fixed by setting the
variable `enable_examples_source_install=false`.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>