qt6-layer: Ignore pyside6 multilib problem [YOCIMX-8902]
For 32-bit multilib config on a 64-bit target:
```
require conf/multilib.conf
MULTILIBS = "multilib:lib32"
DEFAULTTUNE:virtclass-multilib-lib32 = "armv7athf-neon"
```

The 64-bit builds of python3-pyside6 and python3-shiboken6 fail for the
QA test installed-vs-shipped:
```
ERROR: python3-shiboken6-6.8.1-r0 do_package: QA Issue: python3-shiboken6: Files/directories were installed but not shipped in any package:
  /usr/lib
  /usr/lib/python3.13
  /usr/lib/python3.13/site-packages
  /usr/lib/python3.13/site-packages/shiboken6
  /usr/lib/python3.13/site-packages/shiboken6/_git_shiboken_module_version.py
  /usr/lib/python3.13/site-packages/shiboken6/_config.py
  /usr/lib/python3.13/site-packages/shiboken6/Shiboken.pyi
  /usr/lib/python3.13/site-packages/shiboken6/__init__.py
  /usr/lib/python3.13/site-packages/shiboken6/py.typed
  /usr/lib/python3.13/site-packages/shiboken6/Shiboken.abi3.so
```

In this case, the packages should be installing in `/usr/lib64`, not
`/usr/lib`, hence the error.

This commit just fixes the build by ignoring the problem, and likely
just delays the manifestation of the problem until runtime. We do not
test the runtime for the multilib case.

The problem is reported to Qt here:
https://bugreports.qt.io/browse/PYSIDE-3080

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>