parsec-service: Fix cargo error: unexpected argument [LF-14936]
Fix the do_compile error:
```
| NOTE: cargo build -v --frozen --target aarch64-poky-linux-gnu --release --manifest-path=/home/r60874/walnascar/build/tmp/work/armv8a-poky-linux/parsec-service/1.4.1/parsec-service-1.4.1//Cargo.toml  --features mbed-crypto-provider,pkcs11-provider,cryptoki/generate-bindings  mbed-crypto-provider pkcs11-provider cryptoki/generate-bindings
| error: unexpected argument 'mbed-crypto-provider' found
|
| Usage: cargo build [OPTIONS]
```

The issue is discussed on the mailing list, with a proposed fix here:
https://lists.openembedded.org/g/openembedded-core/topic/111812482#msg216762

The problem is cargo.bbclass does now add PACKAGECONFIG_CONFARGS
directly to the cargo command line. This breaks the parsec-service
recipe because it does already add PACKAGECONFIG_CONFARGS to the
command line via CARGO_BUILD_FLAGS.

While this is being fixed upstream, implement the latest proposed fix
via bbappend, adapting the recipe for the new cargo.bbclass behavior.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>