Add oem-prov-app application recipe
The OEM Provisioning application is a tool designed
to facilitate the OEM Provisioning process. It supports
the import of EdgeLock 2GO security assets into the
EdgeLock Enclave.
Also moved the existing recipe into a separate folder
as it is not part of the NXP IOT Agent, but a different
application.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>