imx-image-multimedia: Drop psplash for Wayland images [LF-15785,LF-15619]
psplash provides minimal benefit on Wayland-based images, flashing a
screen 'Yocto Project' for a couple of seconds and a progress bar that
barely moves. What's worse, for images with Xen and psplash, some
unknown problem causes Weston failures, while removing psplash seems
to make the problem go away.

So, just remove psplash to simplify the Wayland images and to avoid
the failure on Xen.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>