smw: Re-enable buildpaths QA [SSMW-966]
The el2go-agent failed to build with the following error:
```
| ninja: error: '/home/bamboo/bamboo-agent-home/xml-data/build-dir/IMXYOCTO-LFML-IMX8ULP/elastic-ci-tools/556/build/tmp/work/armv8a-mx8ulp-poky-linux/smw/git/recipe-sysroot/usr/lib/libsqlite3.so', needed by 'libnxp_iot_agent.so.6.3.0',>
```

The problem is apparent in the path shown, which points to the smw
working directory that is only valid for the smw build. The incorrect
path was provided by smw in smw-dev cmake file. This is now fixed.

The issue was in fact detected by Yocto QA error 'buildpaths', but the
error was suppressed. Now that smw-dbg and smw-dev packages are fixed,
the QA check is re-enabled. Note that smw-tests does still have the
QA error, and the error remains suppressed for that case.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>