psplash: Split start service customization [LF-15979]
For 7D Sabre, the splash screen blinks on and off 3 times while booting.
Removing the psplash --no-console-switch removes this behavior. That
flag was added for DRM-based SOCs, where weston is updated to use seatd,
and is likely not needed at all for FBDEV-based SOCs.

That's half of the customizations in the psplash-start drop-in imx.conf.
The other half is for disabling psplash start for EPDC devices or for
'foreground' devices. This is not needed at all for DRM-bases SOCs.

Split the drop-in, one for i.MX 6 and 7, and one for the rest.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
(cherry picked from commit 6d1a906d65f344bc1d604318ce787a166d44964b)