bootloader: Re-work DDR ECC support [YOCIMX-9331]
For 943 and 95, add the UBOOT_CONFIG `sd-ecc` so a bootloader with DDR
ECC can be selected.

For imx-oei, drop the package config `ddr`. The recipe will always
support 'DDR', so presenting it as an option is confusing. Add instead
`ecc`, which is set by default if `UBOOT_CONFIG` contains `sd-ecc`.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>