imx-gpu-viv,mali-imx: Normalize Vulkan ICD file [YOCIMX-9599]
The Vulkan ICD files for Vivante and Mali are installed to the
filesystem hierarchy /etc, which is incorrect. Install to the hierarchy
/usr/share instead.

For Vivante, the file is renamed from imx_icd.json to
verisilicon_icd.json. The upstream task do_install refers to the file
directly, so add prepend and append operations to adapt for the new
name and location.

For Mali, the file is renamed from icd_mali.json to mali_icd.json.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>